✨ develop the getAddress function
This function is used to determine the future address of an account given
its login hash. It implement the formula used by the CREATE2 opcode.

This commit installs a new dependency: OpenZeppelin Contracts.
Some of the utilitary smart-contracts developed by the openzeppelin team
will be used for the development of our smart-contracts. As OZ's team
forces us to use the version 0.8.20 of the compiler, I had to bump our
version of the compiler to this version.