✨ develop the checkAccountExistence function
This function is used to check if an account already exists for a given
login hash. It either returns the address(0) if the account does not
exist, or the address of the account if it does.
For testing purposes, I created a wrapper contract that exposes the
internal functions of the factory contract as a public function.