✨ develop the signature recovery function
This function is used to check if the name service signature is signed
by the authorized entity. The function reconstructs EIP-191 message hash
based on the provided message (the login hash) then recovers the
signer's address from the signature. The function returns a boolean.
As it is an internal function, it has been added to the test wrapper.