✨ develop the one-step create account flow
This commit implements the function to create and initializing an
account. This function can be used in a one-step flow
where the account is created and directly initialized with a signer.
In addition to the unit tests, a new integration test has been added
to garantee the deterministic deployment of the account matches the
formulas defined in the getAddress function and both flow works the same.
It's important to note this function also deploy a ERC1967Proxy for the
user using the provided login hash. The additionnal arguments are used to
initialize the proxy with the provided signer.
A script has been written to call the function from the CLI.
Last but not least, it is important to mention this function can revert
either by the recovery library with use or by the factory itself when the
signature provided is valid but the signer is not the one expected.

Some notes have been added at the end of the file to give some details
about the factory contract and the natspec documentation of the contract
has been written.