🔨 make BaseScript more generic/powerful
This light refactoring of BaseScript makes it more generic and powerful.
Based on the flag you provide when running the script, it will either
start the hardware wallet flow or use the private key you provide.

In addition, I added new comments for each script to detail how to use
them.