👷 Temporary disable Slither's workflow
Since the PR #18, Slither throws an error related to IR when analyzing
our repository. I do not know where this issue comes from and I have the
feeling the issue is in Slither's hands. That's why I posted a comment to
their repository
[here](https://github.com/crytic/slither/issues/2217#issuecomment-1889667178)

Until we find the fix, I'm disabling the CI workflow that runs Slither.