♻️ factory: switch from recover() to tryRecover()
Instead of using `recover()` that throws an exception when the signature
is invalid, we use `tryRecover()` that returns the recovered address and
the potential error encountered during the recovery process.
It is better to control the error handling and avoid the use of exceptions
in the code especially for a factory that is not supposed to revert.