✨ paymaster: split admin to owner/operator role
Split the owner into two different roles: owner and operator.
- The owner is the role that can withdraw the funds from the paymaster and
 manage its stake.
- The operator is the expected signer of the paymaster signature. It is the
 role in charge of the paymaster's operation.
Both roles can be the same address and can update the operator address.
Only the owner can transfer its role or renounce it.

Additionnaly, the paymaster now has a versionning system to ensure the
compatibility of the paymaster with the rest of the system.