🏗️ restructure the repository and add versionning
Move the contracts to a v1 directory, the generic utils
contracts to the utils directory and rename `Account.sol`
to `SmartAccount.sol`. `Account` is a already defined
object in Foundry, leading to some deps collisions.
Also, for each version of the protocol, we now have
a metadata library that will be used to store
the metadata of the protocol such as the version of
it. This commit sets the version for each contracts of
the protocol (SmartAccount, AccountFactory, Paymaster)