✨ account: manage WAR1 signer addition
This commit adds the ability to add another WAR1
signer to the account. To prevent overriding an
existing signer, the function will revert if a
signer already exist with the same cred id hash.

This function can only be called by the contract
itself, meaning the whole flow defined in the
EIP-4337 must be respected (including the signature
validation).